ai-service/requirements.txt
# AI Service Dependencies for 2nd-Brain-Auto
# Core AI and ML libraries
openai>=1.0.0
anthropic>=0.40.0
httpx>=0.24.0
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0
wordcloud>=1.9.0
gensim>=4.3.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx>=0.24.0

# Data processing and storage
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.65.0

# Optional: Advanced AI features
langchain>=0.0.200
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
